refactor: move component props from separated file to component file
