--- src/components/UI/itemProperty/itemPropertyProps.ts
@@ -1,4 +1,0 @@
-export type ItemPropertyProps = {
-  text: string;
-  value?: string | number;
-};
