--- src/components/UI/itemRatingProperty/itemRatingPropertyProps.ts
@@ -1,6 +1,0 @@
-import { BaseProps } from '../../../types/baseProps';
-
-export type ItemRatingPropertyProps = {
-  rating?: number;
-  voted?: number;
-} & BaseProps;
