+++ src/components/UI/itemsControl/itemsControl.tsx
@@ -1,8 +1,12 @@
 import { Pagination } from './pagination/pagination';
 import classNames from './itemsControl.module.css';
-import { ItemsControlProps } from './itemsControlProps';
 import { TotalCount } from './totalCount/totalCount';
 import { PageSizeForm } from './pageSizeForm/pageSizeForm';
+import { TVShowResultsConfig } from '../../../types/api/apiResults';
+
+export type ItemsControlProps = {
+  resultsConfig: TVShowResultsConfig;
+};
 
 export function ItemsControl(props: ItemsControlProps) {
   return (
