--- src/components/UI/itemsControl/itemsControlProps.ts
@@ -1,5 +1,0 @@
-import { TVShowResultsConfig } from '../../../types/api/apiResults';
-
-export type ItemsControlProps = {
-  resultsConfig: TVShowResultsConfig;
-};
