--- src/components/UI/itemsControl/pageSizeForm/pageSizeFormProps.ts
@@ -1,3 +1,0 @@
-export type PageSizeFormProps = {
-  pageSize: number;
-};
