--- src/components/UI/itemsControl/pagination/paginationButton/paginationButtonProps.ts
@@ -1,8 +1,0 @@
-import { BaseProps } from '../../../../../types/baseProps';
-
-export type PaginationButtonProps = {
-  pageNumber?: number;
-  current?: boolean;
-  disabled: boolean;
-  onClick: (pageNumber: number) => void;
-} & BaseProps;
