--- src/components/UI/itemsControl/pagination/paginationProps.ts
@@ -1,5 +1,0 @@
-import { TVShowResultsConfig } from '../../../../types/api/apiResults';
-
-export type PaginationProps = {
-  resultsConfig: TVShowResultsConfig;
-};
