--- src/components/UI/itemsControl/totalCount/totalCountProps.ts
@@ -1,3 +1,0 @@
-export type TotalCountProps = {
-  count: number;
-};
