+++ src/components/UI/loader/loader.tsx
@@ -1,5 +1,10 @@
 import classNames from './loader.module.css';
-import { LoaderProps } from './loaderProps';
+
+type LoaderProps = {
+  width?: number;
+  height?: number;
+};
+
 export function Loader({ width = 48, height = 48 }: LoaderProps) {
   return <div style={{ width: width, height: height }} className={classNames.itemsLoader} />;
 }
