--- src/components/UI/loader/loaderProps.ts
@@ -1,4 +1,0 @@
-export type LoaderProps = {
-  width?: number;
-  height?: number;
-};
