--- src/components/app/appProps.ts
@@ -1,3 +1,0 @@
-import { BaseProps } from '../../types/baseProps';
-
-export type AppProps = BaseProps;
