--- src/components/error/errorBoundary/errorBoundaryState.ts
@@ -1,4 +1,0 @@
-export type ErrorBoundaryState = {
-  error?: Error;
-  errorInfo?: React.ErrorInfo;
-};
