--- src/components/error/errorFallback/errorFallbackProps.ts
@@ -1,4 +1,0 @@
-export type ErrorFallbackProps = {
-  error?: Error;
-  errorInfo?: React.ErrorInfo;
-};
