--- src/components/tvShow/tvShowCard/tvShowCardProps.ts
@@ -1,5 +1,0 @@
-import { TVShow } from '../../../types/api/tvShow';
-
-export type CardProps = {
-  item: TVShow;
-};
