+++ src/components/tvShow/tvShowList/tvShowList.tsx
@@ -1,6 +1,11 @@
 import { Card } from '../tvShowCard/tvShowCard';
-import { TVShowListProps } from './tvShowListProps';
 import classNames from './tvShowList.module.css';
+import { TVShow } from '../../../types/api/tvShow';
+
+export type TVShowListProps = {
+  items: Array<TVShow>;
+};
+
 export function TVShowList(props: TVShowListProps) {
   return (
     <section className={classNames.itemsWrapper}>
