--- src/components/tvShow/tvShowList/tvShowListProps.ts
@@ -1,5 +1,0 @@
-import { TVShow } from '../../../types/api/tvShow';
-
-export type TVShowListProps = {
-  items: Array<TVShow>;
-};
