--- src/components/tvShow/tvShows/tvShowsProps.ts
@@ -1,4 +1,0 @@
-import { TVShowResults } from '../../../types/api/apiResults';
-import { SearchFnProps } from '../../../types/searchFunctionProps';
-
-export type TVShowsProps = TVShowResults & SearchFnProps;
